requirements.txt
langchain
langchain-community
langchain-groq
langchain-core
langgraph
chromadb
streamlit
sentence-transformers==2.7.0
python-dotenv
requests
plotly
pandas
pysqlite3-binary
